Update chapter-05 exercise 28
I guess my solution much more descriptive. And i think it's a little bit confusing for readers like me that studying chapter 5 and don't know about assembly yet to see the answers only in assembly without machine code (which fifth chapter dedicated to). Moreover, looks like there is mistake here: 
1. LD R1 #2 
2. STR R0 R1 #0 
The number of instruction changed to +1, so the xF3FF value that you want to load in R1 is now not in the x3004 like in the original code, but in the x3005 instead. Check my version to see what I'm talking about.