Update chapter-05 exercise 25
You can't just halt the program in BRz case, because zero in this case assigned to R5 rather than R1, as it should be by conditions of exercise. Again, it's a little bit strange to see only assembly code before the assembly chapter, so I propose my descriptive and readable table.