Update chapter-06 exercise 21
That is the mistake in the textbook answers. It's can't be R0+1 since this is the case when x3101 is 0 ! More over we already have at x3005 R0 = R0 + 1 that correspond to case when buton up pressed and x3101 set to 1. So turns out in case of textbook answer we up volume in any case :)